+++ library/alloctests/tests/io/buffered.rs
@@ -1,10 +1,10 @@
-use crate::io::prelude::*;
-use crate::io::{
+use alloc::io::prelude::*;
+use alloc::io::{
     self, BorrowedBuf, BufReader, BufWriter, ErrorKind, IoSlice, LineWriter, SeekFrom,
 };
-use crate::mem::MaybeUninit;
-use crate::sync::atomic::{AtomicUsize, Ordering};
-use crate::{panic, thread};
+use core::mem::MaybeUninit;
+use core::sync::atomic::{AtomicUsize, Ordering};
+use std::{panic, thread};
 
 /// A dummy reader intended at testing short-reads propagation.
 pub struct ShortReader {
@@ -515,29 +515,6 @@ fn panic_in_write_doesnt_flush_in_drop() {
     assert_eq!(WRITES.load(Ordering::SeqCst), 1);
 }
 
-#[bench]
-fn bench_buffered_reader(b: &mut test::Bencher) {
-    b.iter(|| BufReader::new(io::empty()));
-}
-
-#[bench]
-fn bench_buffered_reader_small_reads(b: &mut test::Bencher) {
-    let data = (0..u8::MAX).cycle().take(1024 * 4).collect::<Vec<_>>();
-    b.iter(|| {
-        let mut reader = BufReader::new(&data[..]);
-        let mut buf = [0u8; 4];
-        for _ in 0..1024 {
-            reader.read_exact(&mut buf).unwrap();
-            core::hint::black_box(&buf);
-        }
-    });
-}
-
-#[bench]
-fn bench_buffered_writer(b: &mut test::Bencher) {
-    b.iter(|| BufWriter::new(io::sink()));
-}
-
 /// A simple `Write` target, designed to be wrapped by `LineWriter` /
 /// `BufWriter` / etc, that can have its `write` & `flush` behavior
 /// configured
@@ -681,7 +658,7 @@ fn line_vectored() {
 
 #[test]
 fn line_vectored_partial_and_errors() {
-    use crate::collections::VecDeque;
+    use alloc::collections::VecDeque;
 
     enum Call {
         Write { inputs: Vec<&'static [u8]>, output: io::Result<usize> },
@@ -1023,7 +1000,7 @@ struct WriteRecorder {
 
 impl Write for WriteRecorder {
     fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        use crate::str::from_utf8;
+        use core::str::from_utf8;
 
         self.events.push(RecordedEvent::Write(from_utf8(buf).unwrap().to_string()));
         Ok(buf.len())
@@ -1056,7 +1033,7 @@ fn single_formatted_write() {
 fn bufreader_full_initialize() {
     struct OneByteReader;
     impl Read for OneByteReader {
-        fn read(&mut self, buf: &mut [u8]) -> crate::io::Result<usize> {
+        fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
             if buf.len() > 0 {
                 buf[0] = 0;
                 Ok(1)
@@ -1079,9 +1056,8 @@ fn bufreader_full_initialize() {
 /// This is a regression test for https://github.com/rust-lang/rust/issues/127584.
 #[test]
 fn bufwriter_aliasing() {
-    use crate::io::{BufWriter, Cursor};
     let mut v = vec![0; 1024];
-    let c = Cursor::new(&mut v);
-    let w = BufWriter::new(Box::new(c));
+    let c = io::Cursor::new(&mut v);
+    let w = io::BufWriter::new(Box::new(c));
     let _ = w.into_parts();
 }
