+++ library/std/src/io/mod.rs
@@ -294,9 +294,6 @@
 
 #![stable(feature = "rust1", since = "1.0.0")]
 
-#[cfg(test)]
-mod tests;
-
 #[unstable(feature = "raw_os_error_ty", issue = "107792")]
 pub use alloc::io::RawOsError;
 #[doc(hidden)]
@@ -345,3 +342,6 @@ pub(crate) use alloc::io::{
 };
 
 pub(crate) use stdio::cleanup;
+
+#[cfg(test)]
+mod io_benches;
